requirements.txt
streamlit
requests
numpy
pillow
google-generativeai
opencv-python-headless
pyzbar
